split e2e tests into chunks
