+++ packages/python/package.json
@@ -17,7 +17,7 @@
     "build": "node build",
     "test": "jest --env node --verbose --runInBand --bail",
     "test-unit": "pnpm test test/unit.test.ts",
-    "test-e2e": "pnpm test test/integration.test.ts"
+    "test-e2e": "pnpm test test/integration-*"
   },
   "devDependencies": {
     "@types/execa": "^0.9.0",
