+++ packages/python/test/integration-1.test.ts
@@ -71,30 +71,11 @@ it('should match the probes against Python dev servers', async () => {
   }
 });
 
-const testsThatFailToBuild = new Map([
-  ['30-fail-build-invalid-pipfile', 'Unable to parse Pipfile.lock'],
-  [
-    '31-fail-build-invalid-python36',
-    'Python version "3.6" detected in Pipfile.lock is discontinued and must be upgraded.',
-  ],
-]);
+const allFixtures = fs.readdirSync(fixturesPath);
+const chunkFixtures = allFixtures.slice(0, 10);
 
 // eslint-disable-next-line no-restricted-syntax
-for (const fixture of fs.readdirSync(fixturesPath)) {
-  const errMsg = testsThatFailToBuild.get(fixture);
-  if (errMsg) {
-    // eslint-disable-next-line no-loop-func
-    it(`should fail to build ${fixture}`, async () => {
-      try {
-        await testDeployment(path.join(fixturesPath, fixture));
-      } catch (err) {
-        expect(err).toBeTruthy();
-        expect(err.deployment).toBeTruthy();
-        expect(err.deployment.errorMessage).toBe(errMsg);
-      }
-    });
-    continue; //eslint-disable-line
-  }
+for (const fixture of chunkFixtures) {
   // eslint-disable-next-line no-loop-func
   it(`should build ${fixture}`, async () => {
     await expect(
